Omit trivial variables from created MPS files
In the tofile() method of the op class we ensure a variable is not
trivial before inserting an entry for it in the BOUNDS section of the
resulting fixed MPS file. We define a variable as trivial if all
coefficients for that variable are 0 in the objective and 0 in all of
the constraints. Note CVXOPT already omits these variables from
appearing in the COLUMNS section of the fixed MPS file; this change
makes it so they are omitted from the BOUNDS section as well.

The fixed MPS file format conventions don't seem to necessarily
prohibit having trivial variables in the BOUNDS section, however;
solvers such as clp (COIN-OR) and glpsol (GLPK) throw an error in such
cases with regards to being unable to find a variable if it appears in
the BOUNDS section but does not appear in the COLUMNS section.