REL: pass parameter `python_requires` to `setuptools.setup`
in the script `setup.py`, to specify the Python versions
that `dd` requires.

The parameter `python_requires` defines the metadata field
`Requires-Python` that is specified in [PEP 345](
    https://www.python.org/dev/peps/pep-0345/#requires-python).
[PEP 503](https://www.python.org/dev/peps/pep-0503/#specification)
specifies how installer tools behave when `Requires-Python` is
present and communicated by PyPI.

Users need to use [`pip >= 9.0.0`](
    https://pip.pypa.io/en/stable/news/#v9-0-0)
and [`setuptools >= 24.3.0`](
    https://setuptools.readthedocs.io/en/latest/history.html#v24-3-0)
for the parameter `python_requires` to have the intended effect.

Relevant resources:
- https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires
- https://packaging.python.org/guides/dropping-older-python-versions/
- https://packaging.python.org/guides/distributing-packages-using-setuptools/#classifiers